Fix a bug in add_faux_cell
This line checks that the row doesn't already exist in the gridmap
array; if it does then it doesn't need to be pushed to the faux_grid
array.

Otherwise, this function introduced a bug when I was dynamically adding
widgets that had a row height > 1 as the first widget added. Any
subsequent widget added would end up being at row 2 (rather than row 3,
as the first widget already occupied the first 2 rows).